backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-environ==0.8.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-cors-headers==3.10.1
django-environ==0.8.1
django-mapbox-location-field==1.7.3
django-resized==0.3.11
django-tinymce==3.4.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-cors-headers==3.10.1
django-environ==0.8.1
django-mapbox-location-field==1.7.3
django-resized==0.3.11
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
